refactor: seed-test
