+++ package.json
@@ -25,7 +25,7 @@
     "prisma:generate": "npx prisma generate",
     "prisma:studio": "npx prisma studio",
     "prisma:seed": "npx prisma db seed",
-    "prisma:seed:test": "SEED_TEST_DATA=1 npx prisma db seed",
+    "prisma:seed:test": "ts-node prisma/seed-test.ts",
     "prepare": "husky",
     "commit": "git-cz",
     "release": "standard-version"
